+++ conf/run.sh
@@ -24,12 +24,7 @@ case "$1" in
             miracle $2 $3 $4 $5 $6 $7 $8 $9 && \
             flake8 miracle
         ;;
-    alembic)
-        echo "Running Alembic"
-        cd ..
-        alembic $2 $3 $4 $5 $6 $7 $8 $9
-        ;;
     *)
-        echo "Usage: $0 {web|worker|shell|test|alembic}"
+        echo "Usage: $0 {web|worker|shell|test}"
         exit 1
 esac